Replaces final sitemaps with static alternatives
Finishes sitemap project by adding lifestream, portfolio, and
waterfall sitemaps to the generation script. Several new data
repositories had to be added for this to occur with some basic
methods (note: they will have to be revisited). Also removed
the old sitemap controller and associated stack.

Note: this script is not in its final form, just a step foward.